id로 select 테스트 및 예외처리

+++ src/main/java/com/cos/blog/model/User.java
@@ -21,7 +21,7 @@ public class User {
 
     @Id //Primary key
     @GeneratedValue(strategy = GenerationType.IDENTITY)//프로젝트에서 연결된 DB의 넘버링 전략을 따라간다.
-    private Long id; // 시퀀스, auto_increment
+    private int id; // 시퀀스, auto_increment
 
     @Column(nullable = false, length = 30)
     private String username; //아이디
